requirements.txt
setuptools==51.1.0
wheel==0.36.2
pyyaml==5.4.1
pandas==1.2.4